requirements.txt
anyio==3.7.0
atomicwrites==1.4.1
attrs==23.1.0
bleach==6.0.0
certifi==2023.5.7
chardet==4.0.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.1.1
fastapi==0.96.0
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
invoke==1.7.3
jaraco.classes==3.2.3
keyring==23.13.1
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
pluggy==1.0.0
postgrest==0.10.6
postgrest-py==0.10.6
py==1.11.0
pydantic==1.10.9
Pygments==2.15.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.14.0
python-semantic-release==7.33.2
pywin32-ctypes==0.2.0
readme-renderer==37.3
realtime==1.0.0
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
toml==0.10.2
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
webencodings==0.5.1
websockets==11.0.3
zipp==3.15.0
